==> notebooks/06_motifX.ipynb <==
# PO-4

Section 6. Identification of new motifs with MotifX 

## Table of Contents <a class="anchor" id="table-of-contents"></a>

* [1. Introduction](#introduction)        
 
* [2. Loading foreground and background motifs](#data-loading)    

* [3. MotifX analysis](#motifx)
    * [3.1 Serine 13-mers](#serine)
    * [3.2 Threonine 13-mers](#threonine)
    * [3.3 Tyrosine 13-mers](#tyrosine)

# 1. Introduction <a class="anchor" id="introduction"></a>

https://motif-x.med.harvard.edu/faq.html is not running now. 

Therefore we have used the R package implemented in https://github.com/omarwagih/rmotifx

install.packages('devtools')
require(devtools)

install_github('omarwagih/rmotifx')
require(rmotifx)

?motifx

# 2. Loading foreground and background motifs <a class="anchor" id="data-loading"></a>

fgs.seqs <- readLines('../data/interim/serine_fg.tsv')
bgs.seqs <- readLines('../data/interim/serine_bg.tsv')
serines_mot = motifx(fgs.seqs, bgs.seqs, central.res = 'S', min.seqs = 5, pval.cutoff = 1e-3)

fgt.seqs <- readLines('../data/interim/threonine_fg.tsv')
bgt.seqs <- readLines('../data/interim/threonine_bg.tsv')
threonine_mot = motifx(fgt.seqs, bgt.seqs, central.res = 'T', min.seqs = 5, pval.cutoff = 1e-3)

fgy.seqs <- readLines('../data/interim/tyrosine_fg.tsv')
bgy.seqs <- readLines('../data/interim/tyrosine_bg.tsv')
tyrosine_mot = motifx(fgy.seqs, bgy.seqs, central.res = 'Y', min.seqs = 5, pval.cutoff = 1e-3)

# 3. MotifX analysis <a class="anchor" id="motifx"></a>

## 3.1 Serine 13-mers <a class="anchor" id="serine"></a>

length(fgs.seqs)
head(fgs.seqs)
length(bgs.seqs)
head(bgs.seqs)

print(serines_mot)

## 3.2 Threonine 13-mers <a class="anchor" id="threonine"></a>

length(fgt.seqs)
head(fgt.seqs)
length(bgt.seqs)
head(bgt.seqs)

print(threonine_mot)

## 3.3 Tyrosine 13-mers <a class="anchor" id="tyrosine"></a>

length(fgy.seqs)
head(fgy.seqs)
length(bgy.seqs)
head(bgy.seqs)

print(tyrosine_mot)
